Altimeter: fix output format for hPa/InHg settings
Usually, in a real cockpit, altimeter shows hPa setting as an integer number and InHg setting is limited to 2 decimals

In aviation we don't need higher numeric precision

+++ src/Instrumentation/altimeter.hxx
@@ -43,8 +43,8 @@ public:
 
     double getSettingInHg() const;
     void setSettingInHg( double value );
-    double getSettingHPa() const;
-    void setSettingHPa( double value );
+    int getSettingHPa() const;
+    void setSettingHPa( int value );
 
 private:
     std::string _name;
